+++ tests/test_config.py
@@ -177,6 +177,13 @@ def test_get_all_connections(test_snowcli_config):
             "private_key_file": "/private/key",
             "user": "jdoe",
         },
+        "wif": {
+            "workload_identity_provider": "GCP",
+        },
+        "wif_oidc": {
+            "workload_identity_provider": "OIDC",
+            "token": "my-oidc-config-token",
+        },
     }